+++ .storybook/config.js
@@ -1,6 +1,6 @@
 import { configure } from '@storybook/react';
 
-const requireContext = require.context('../source', true, /\.story\.js$/)
+const requireContext = require.context('../source', true, /(\/|\.)story\.js$/);
 
 function loadStories() {
   requireContext.keys().forEach(requireContext);