+++ source/badge/index.js
@@ -1,12 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
-import styles from './badge.css';
+import styles from './styles.css';
 
 const Badge = ({ className, children }) => (
-  <span className={cn(className, styles.root)}>
-    {children}
-  </span>
+  <span className={cn(className, styles.root)}>{children}</span>
 );
 
 Badge.propTypes = {