+++ source/badge/story.js
@@ -2,9 +2,4 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import Badge from './';
 
-storiesOf('Badge', module)
-  .add('default', () => (
-    <Badge>
-      Badge
-    </Badge>
-  ));
+storiesOf('Badge', module).add('default', () => <Badge>Badge</Badge>);