+++ source/meter/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
-import styles from './meter.css';
+import styles from './styles.css';
 
 export default function Meter({ className, value }) {
   return (