+++ source/select/index.js
@@ -1,24 +1,32 @@
-import React from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
-import styles from './select.css';
+import FieldError from '../field-error';
+import styles from './styles.css';
 
-export default function Select({ className, children, ...restProps }) {
-  return (
-    <div className={cn(className, styles.root)}>
-      <select {...restProps}>
-        {children}
-      </select>
-    </div>
-  );
+export default class Select extends Component {
+  render = () => {
+    const { className, children, error, ...restProps } = this.props;
+
+    return (
+      <div
+        className={cn(className, styles.root, {
+          [styles.root_invalid]: Boolean(error),
+        })}
+      >
+        <div className={styles['select-wrapper']}>
+          <select className={styles.select} {...restProps}>
+            {children}
+          </select>
+        </div>
+        <FieldError>{error}</FieldError>
+      </div>
+    );
+  };
 }
 
 Select.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node,
-};
-
-Select.defaultProps = {
-  className: '',
-  children: null,
+  error: PropTypes.string,
 };